Split StoreUtil out into StoreInitializer and StoreDefault
