+++ store/src/main/java/com/nytimes/android/external/store/base/impl/StoreDefaults.java
@@ -1,0 +1,31 @@
+package com.nytimes.android.external.store.base.impl;
+
+import java.util.concurrent.TimeUnit;
+
+final class StoreDefaults {
+    private StoreDefaults() {
+
+    }
+
+    /**
+     * Default Cache TTL, can be overridden
+     *
+     * @return memory persister ttl
+     */
+    static long getCacheTTL() {
+        return TimeUnit.HOURS.toSeconds(24);
+    }
+
+    /**
+     * Default mem persister is 1, can be overridden otherwise
+     *
+     * @return memory persister size
+     */
+    static long getCacheSize() {
+        return 100;
+    }
+
+    static TimeUnit getCacheTTLTimeUnit() {
+        return TimeUnit.SECONDS;
+    }
+}
